feat: Add documentation and update Reminder data class
Added documentation to the `Reminder` data class explaining its properties.
Updated the `id` property type from
 `Long` to `Int`.
This change likely aligns with how reminders are identified.

+++ core/local/repository/src/main/java/city/zouitel/repository/model/Reminder.kt
@@ -2,9 +2,17 @@ package city.zouitel.repository.model
 
 import androidx.annotation.Keep
 
+/**
+ * Data class representing a reminder.
+ *
+ * @property id The unique identifier of the reminder. Defaults to 0.
+ * @property uid The unique identifier of the user this reminder belongs to. Defaults to an empty string.
+ * @property atTime The time when the reminder is scheduled, in milliseconds since the epoch. Defaults to 0L.
+ * @property isPassed A flag indicating whether the reminder has already passed its scheduled time. Defaults to false.
+ */
 @Keep
 data class Reminder(
-    val id: Long = 0L,
+    val id: Int = 0,
     val uid: String = "",
     val atTime: Long = 0L,
     val isPassed: Boolean = false
